requirements.txt
allure-behave==2.13.2
allure-python-commons==2.13.2
attrs==23.2.0
behave==1.2.6
certifi==2023.11.17
Faker==23.1.0
h11==0.14.0
idna==3.6
outcome==1.3.0.post0
parse==1.20.0
parse-type==0.6.2
pluggy==1.4.0
PySocks==1.7.1
python-dateutil==2.8.2
selenium==4.16.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.24.0
trio-websocket==0.11.1
urllib3==2.1.0
wsproto==1.2.0
